Корректирует di по модулям
